Move operation serialization to MetadataSplitter
